+++ examples/react-hooks/e-commerce/App.tsx
@@ -249,6 +249,7 @@ function SubmitIcon() {
       width="16"
       height="16"
       viewBox="0 0 18 18"
+      aria-hidden="true"
     >
       <g
         fill="none"