+++ packages/instantsearch.js/src/connectors/hierarchical-menu/connectHierarchicalMenu.ts
@@ -411,6 +411,19 @@ const connectHierarchicalMenu: HierarchicalMenuConnector =
             uiState.hierarchicalMenu &&
             uiState.hierarchicalMenu[hierarchicalFacetName];
 
+          if (
+            searchParameters.isConjunctiveFacet(hierarchicalFacetName) ||
+            searchParameters.isDisjunctiveFacet(hierarchicalFacetName)
+          ) {
+            warning(
+              false,
+              `HierarchicalMenu: Attribute "${hierarchicalFacetName}" is already used by another widget applying conjunctive or disjunctive faceting.
+As this is not supported, please make sure to remove this other widget or this HierarchicalMenu widget will not work at all.`
+            );
+
+            return searchParameters;
+          }
+
           if (searchParameters.isHierarchicalFacet(hierarchicalFacetName)) {
             const facet = searchParameters.getHierarchicalFacetByName(
               hierarchicalFacetName