+++ packages/instantsearch.js/src/widgets/menu-select/__tests__/menu-select-test.ts
@@ -238,7 +238,6 @@ describe('menuSelect', () => {
       expect(search.helper!.state).toEqual(
         new SearchParameters({
           index: 'test',
-          clickAnalytics: true,
         })
       );
 
@@ -249,7 +248,6 @@ describe('menuSelect', () => {
       expect(search.helper!.state).toEqual(
         new SearchParameters({
           index: 'test',
-          clickAnalytics: true,
           hierarchicalFacets: [{ attributes: ['test'], name: 'test' }],
           hierarchicalFacetsRefinements: { test: [] },
           maxValuesPerFacet: 10,
@@ -263,7 +261,7 @@ describe('menuSelect', () => {
       expect(render).toHaveBeenCalledTimes(2);
       expect(render).toHaveBeenLastCalledWith(null, container);
       expect(search.helper!.state).toEqual(
-        new SearchParameters({ index: 'test', clickAnalytics: true })
+        new SearchParameters({ index: 'test' })
       );
     });
   });