+++ packages/react-instantsearch-hooks-web/src/ui/__tests__/RefinementList.test.tsx
@@ -267,6 +267,7 @@ describe('RefinementList', () => {
                   type="submit"
                 >
                   <svg
+                    aria-hidden="true"
                     class="ais-SearchBox-submitIcon"
                     height="10"
                     viewBox="0 0 40 40"
@@ -284,6 +285,7 @@ describe('RefinementList', () => {
                   type="reset"
                 >
                   <svg
+                    aria-hidden="true"
                     class="ais-SearchBox-resetIcon"
                     height="10"
                     viewBox="0 0 20 20"
@@ -299,6 +301,7 @@ describe('RefinementList', () => {
                   hidden=""
                 >
                   <svg
+                    aria-hidden="true"
                     class="ais-SearchBox-loadingIcon"
                     height="16"
                     stroke="#444"