+++ packages/react-instantsearch-hooks-web/src/widgets/__tests__/InfiniteHits.test.tsx
@@ -202,7 +202,6 @@ describe('InfiniteHits', () => {
       {
         indexName: 'indexName',
         params: {
-          clickAnalytics: true,
           facets: [],
           highlightPostTag: '__/ais-highlight__',
           highlightPreTag: '__ais-highlight__',
@@ -224,7 +223,6 @@ describe('InfiniteHits', () => {
         {
           indexName: 'indexName',
           params: {
-            clickAnalytics: true,
             facets: [],
             highlightPostTag: '__/ais-highlight__',
             highlightPreTag: '__ais-highlight__',
@@ -260,7 +258,6 @@ describe('InfiniteHits', () => {
       {
         indexName: 'indexName',
         params: {
-          clickAnalytics: true,
           facets: [],
           highlightPostTag: '__/ais-highlight__',
           highlightPreTag: '__ais-highlight__',
@@ -284,7 +281,6 @@ describe('InfiniteHits', () => {
         {
           indexName: 'indexName',
           params: {
-            clickAnalytics: true,
             facets: [],
             highlightPostTag: '__/ais-highlight__',
             highlightPreTag: '__ais-highlight__',