+++ packages/react-instantsearch-hooks-web/src/widgets/__tests__/SearchBox.test.tsx
@@ -45,6 +45,7 @@ describe('SearchBox', () => {
               type="submit"
             >
               <svg
+                aria-hidden="true"
                 class="ais-SearchBox-submitIcon"
                 height="10"
                 viewBox="0 0 40 40"
@@ -62,6 +63,7 @@ describe('SearchBox', () => {
               type="reset"
             >
               <svg
+                aria-hidden="true"
                 class="ais-SearchBox-resetIcon"
                 height="10"
                 viewBox="0 0 20 20"
@@ -77,6 +79,7 @@ describe('SearchBox', () => {
               hidden=""
             >
               <svg
+                aria-hidden="true"
                 class="ais-SearchBox-loadingIcon"
                 height="16"
                 stroke="#444"