+++ packages/react-instantsearch-hooks-web/src/widgets/__tests__/ToggleRefinement.test.tsx
@@ -102,7 +102,6 @@ describe('ToggleRefinement', () => {
         expect.arrayContaining([
           expect.objectContaining({
             params: {
-              clickAnalytics: true,
               facetFilters: [['free_shipping:true']],
               facets: ['free_shipping'],
               tagFilters: '',
@@ -157,7 +156,6 @@ describe('ToggleRefinement', () => {
         expect.arrayContaining([
           expect.objectContaining({
             params: {
-              clickAnalytics: true,
               facetFilters: [['free_shipping:yes']],
               facets: ['free_shipping'],
               tagFilters: '',