+++ packages/vue-instantsearch/src/components/SearchInput.vue
@@ -38,6 +38,7 @@
     >
       <slot name="submit-icon">
         <svg
+          aria-hidden="true"
           role="img"
           xmlns="http://www.w3.org/2000/svg"
           width="10"
@@ -64,10 +65,10 @@
     >
       <slot name="reset-icon">
         <svg
+          aria-hidden="true"
           role="img"
           xmlns="http://www.w3.org/2000/svg"
-          width="1em"
-          height="1em"
+          style="width: 1em; height: 1em"
           viewBox="0 0 20 20"
           :class="suit('resetIcon')"
         >