+++ tests/common/widgets/infinite-hits/insights.ts
@@ -22,10 +22,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -133,10 +134,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 25;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -264,10 +266,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -344,10 +347,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -428,10 +432,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -522,10 +527,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);
@@ -602,10 +608,11 @@ export function createInsightsTests(setup: InfiniteHitsSetup, act: Act) {
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = jest.fn();
+      window.aa = Object.assign(jest.fn(), { version: '2.4.0' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
+          insights: true,
           searchClient: createSearchClient({
             search: jest.fn(async (requests) => {
               await wait(delay);