+++ tests/mocks/createInsightsClient.ts
@@ -14,17 +14,32 @@ try {
   delete window.AlgoliaAnalyticsObject;
 } catch (error) {} // eslint-disable-line no-empty
 
-export function createInsights() {
+export function createInsights<TVersion extends string | undefined = '2.4.0'>({
+  forceVersion = '2.4.0',
+}: {
+  forceVersion?: TVersion;
+} = {}) {
   const analytics = mockMethods(
     new AlgoliaAnalytics({
       requestFn: jest.fn(),
     })
   );
-  const insightsClient = jest.fn(getFunctionalInterface(analytics));
+
+  if (forceVersion) {
+    return {
+      analytics,
+      insightsClient: Object.assign(
+        jest.fn(getFunctionalInterface(analytics)),
+        {
+          version: forceVersion,
+        }
+      ),
+    };
+  }
 
   return {
     analytics,
-    insightsClient,
+    insightsClient: jest.fn(getFunctionalInterface(analytics)),
   };
 }