+++ packages/react-instantsearch-hooks-router-nextjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.42.1](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-router-nextjs@6.42.0...react-instantsearch-hooks-router-nextjs@6.42.1) (2023-03-28)
+
+**Note:** Version bump only for package react-instantsearch-hooks-router-nextjs
+
+
+
+
+
+## [6.42.0](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-router-nextjs@6.41.0...react-instantsearch-hooks-router-nextjs@6.42.0) (2023-03-21)
+
+**Note:** Version bump only for package react-instantsearch-hooks-router-nextjs
+
+
+
+
+
 ## [6.41.0](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-router-nextjs@6.40.1...react-instantsearch-hooks-router-nextjs@6.41.0) (2023-03-07)
 
 **Note:** Version bump only for package react-instantsearch-hooks-router-nextjs